requirements.txt
matplotlib==3.1.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
multiprocess==0.70.12.2
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634095647912/work
oauthlib==3.2.0
opencv-python==4.2.0.32
p-tqdm==1.3.3
Pillow==9.0.1
scikit-image==0.19.2
scikit-learn==0.23.2
scipy==1.4.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.8.1
torch==1.6.0
torchvision==0.7.0
tqdm==4.43.0

